+++ MonoBehaviors/Stardust_Mono.cs
@@ -1,4 +1,5 @@
-﻿using RSClasses.Utilities;
+﻿using Photon.Pun;
+using RSClasses.Utilities;
 using Sonigon;
 using System;
 using System.Collections;
@@ -17,16 +18,10 @@ namespace RSClasses.MonoBehaviours
     {
         public Player player;
         public float rotationDirection = 90;
-        private SoundEvent stardustSound;
 
         private void Start()
         {
-            AudioClip stardustAudioClip = RSClasses.assets.LoadAsset<AudioClip>("stardustHit.ogg"); // Load sound effects
-            SoundContainer stardustSoundContainer = ScriptableObject.CreateInstance<SoundContainer>();
-            stardustSoundContainer.setting.volumeIntensityEnable = true;
-            stardustSoundContainer.audioClip[0] = stardustAudioClip;
-            stardustSound = ScriptableObject.CreateInstance<SoundEvent>();
-            stardustSound.soundContainerArray[0] = stardustSoundContainer;
+
         }
 
         private void Update()
@@ -36,10 +31,9 @@ namespace RSClasses.MonoBehaviours
 
         private void FixedUpdate()
         {
-            var hits = Physics2D.OverlapCircleAll(transform.position, 0.1f);
-
             if (player.data.view.IsMine)
             {
+                var hits = Physics2D.OverlapCircleAll(transform.position, 0.1f);
                 foreach (var hit in hits)
                 {
                     var damageable = hit.gameObject.GetComponent<Damagable>();
@@ -48,17 +42,30 @@ namespace RSClasses.MonoBehaviours
                     {
                         Player hitPlayer = ((Player)healthHandler.GetFieldValue("player"));
                         if (hitPlayer == player) continue;
-                        SoundManager.Instance.PlayAtPosition(stardustSound, this.transform, damageable.transform);
-                        if (((Player)healthHandler.GetFieldValue("player")).GetComponent<Block>().blockedThisFrame) { continue; }
+                        SoundManager.Instance.PlayAtPosition(RSClasses.stardustSound, this.transform, damageable.transform);
+                        if (((Player)healthHandler.GetFieldValue("player")).GetComponent<Block>().blockedThisFrame)
+                        {
+                            PhotonNetwork.Destroy(this.gameObject);
+                            continue;
+                        }
                     }
                     if (damageable)
                     {
                         damageable.CallTakeDamage(((Vector2)damageable.transform.position - (Vector2)this.transform.position).normalized * (player.data.GetAdditionalData().cometDamage / 10),
                             (Vector2)this.transform.position, this.gameObject, player);
-                        Destroy(this.gameObject);
+                        PhotonNetwork.Destroy(this.gameObject);
                     }
                 }
             }
+        }
+
+        [PunRPC]
+        public void SetValues(int ownerID, int rotation)
+        {
+            rotationDirection = rotation; // And random rotation directions
+            player = PlayerManager.instance.players.Find(p => p.playerID == ownerID); // Set stardust player and color
+            UnityEngine.Debug.Log(player);
+            GetComponent<SpriteRenderer>().color = player.GetTeamColors().color * 1.75f;
         }
     }
 }